backend/requirements.txt
# Core Dependencies - Alle KOSTENLOS!
python-dotenv==1.0.0

# Web Scraping & HTTP
httpx==0.27.0
aiohttp==3.9.1
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0
selenium==4.15.2

# Data Processing & Validation
pydantic==2.6.1
pandas==2.2.0
pyarrow==15.0.0
numpy>=1.21.0

# Async & Concurrency
asyncio==3.4.3
aiofiles==23.2.1
asyncio-throttle==1.0.2

# Utilities
tldextract==5.1.1
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3
requests-robots==0.0.4

# Caching & Storage
redis==5.0.1
diskcache==5.6.3

# Monitoring & Logging
structlog==23.2.0
rich==13.7.0

# LangChain & AI Integration (simplified)
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-openai==0.0.5
pydantic==2.6.1

# OpenAI API (statt lokale LLMs)
openai==1.12.0

# Firecrawl API
firecrawl-py==0.0.8
requests==2.31.0

# FastAPI für Real-time
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Scheduling für automatisierte Pipeline
schedule==1.2.0

# Web Dashboard
flask==3.0.0

# Geospatial Visualization
folium==0.15.1

# H3 Geospatial Indexing
h3==3.7.6

# Free LLM Integration
ollama==0.1.7

# OpenStack Client
python-openstackclient>=6.2.0
openstacksdk>=1.3.0
python-novaclient>=18.0.0
python-glanceclient>=4.0.0
python-neutronclient>=9.0.0
python-cinderclient>=9.0.0
python-keystoneclient>=5.0.0
PyYAML>=6.0
